+++ lib/login_with_connect.dart
@@ -2,3 +2,4 @@ library;
 
 export 'src/connect_auth_client.dart';
 export 'src/connect_auth_exception.dart';
+export 'src/connect_environment.dart';